pyproject.toml
[project]
name = "recipe-api"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "strawberry-graphql>=0.240",
  "sqlalchemy>=2.0",
  "aiosqlite>=0.20",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "anyio>=4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
